move and rename codeanalysis file to findtodos
