+++ spyder/plugins/editor/utils/findtodos.py
